sign up page fixes
